pyproject.toml
[project]
name = "watchflow"
version = "1.0"
description = ""
authors = [{ name = "San Shibu", email = "scorpiocodex0@gmail.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "typer[all] (>=0.20.1,<0.21.0)",
    "rich (>=14.2.0,<15.0.0)",
    "watchdog (>=6.0.0,<7.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "structlog (>=25.5.0,<26.0.0)",
]

[tool.poetry]
packages = [{ include = "watchflow", from = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "mypy (>=1.19.1,<2.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "ruff (>=0.14.10,<0.15.0)",
    "types-pyyaml (>=6.0.12.20250915,<7.0.0.0)",
]

[project.scripts]
watchflow = "watchflow.cli:app"
